Added User Functionality with users having only their devices

+++ route/api/device.js
@@ -9,14 +9,17 @@ const changeStream = Device.watch().on("change", (change) =>
 );
 
 router.get("/", (req, res, next) => {
-  Device.find({}, { _id: 0, __v: 0 })
+  Device.find({ user: req.headers.userid })
+    .select("-_id -__v")
     .populate("description")
     .exec(function (err, device) {
       if (err) return handleError(err);
-      req.encryptUserData = {
-        device,
-        userData: `${req.session.user.firstName} ${req.session.user.lastName}`,
-      };
+      if (device) {
+        req.encryptUserData = {
+          device,
+          userData: `${req.session.user.firstName} ${req.session.user.lastName}`,
+        };
+      }
       next();
     });
 });
@@ -24,11 +27,13 @@ router.get("/", (req, res, next) => {
 router.post("/add", async (req, res, next) => {
   console.log(req.body);
 
+  const userID = req.body.userID;
   const deviceType = req.body.option;
 
   const retrievedDevice = await DeviceType.findOne({ type: deviceType });
 
   const device = new Device({
+    user: userID,
     ...req.body,
     description: retrievedDevice._id,
     status: false,
@@ -42,7 +47,19 @@ router.post("/add", async (req, res, next) => {
 });
 
 router.post("/remove", async (req, res) => {
+  const userID = req.body.userID;
   const deviceID = req.body.deviceID;
+
+  const userDevices = await Device.find({ user: userID });
+
+  const device = userDevices.find((el) => el.deviceID === deviceID);
+  if (!device) {
+    return res.send("Device Doesn't Exist");
+  }
+  if (device.user.toString() !== userID) {
+    return res.send("User not authorized");
+  }
+
   await Device.findOneAndDelete({ deviceID: deviceID }).exec(function (
     err,
     device
@@ -55,6 +72,16 @@ router.post("/remove", async (req, res) => {
 router.post("/update", async (req, res) => {
   const filter = { deviceID: req.body.deviceID };
   const update = { status: req.body.status };
+
+  const userDevices = await Device.find({ user: req.body.userID });
+  const device = userDevices.find((el) => el.deviceID === req.body.deviceID);
+
+  if (!device) {
+    return res.send("Device Doesn't Exist");
+  }
+  if (device.user.toString() !== req.body.userID) {
+    return res.send("User not authorized");
+  }
 
   await Device.findOneAndUpdate(filter, update).exec(function (err, device) {
     if (err) console.log(err);
